refactor: split backend query and foundation modules
