+++ crates/db/src/queries/common.rs
@@ -1,0 +1,11 @@
+pub(super) use crate::models::{
+    AgentEssence, Avatar, AvatarArtifactTrail, AvatarDebateEvent, AvatarIdentityState,
+    AvatarMemoryEdge, AvatarPresenceState, AvatarSoul, Campaign, CampaignBrief, CampaignMove,
+    CampaignTask, CapabilityArtifact, CapabilityDefinition, CapabilityRun, CompetitorSnapshot,
+    ContentStrategy, CouncilAgentPosition, CouncilAvatarTurn, CouncilOrchestrationRun,
+    CouncilSession, DailyWin, FoundationScan, FoundationSection, FoundationSnapshot,
+    FoundationVersion, GeneratedContent, HarnessContextPack, HarnessRun, HarnessStep,
+    MuseConversation, MuseMessage, Nudge, OrgUser, Organization, ReplanSession, Ripple, RippleEdge,
+    Subscription,
+};
+pub(super) use sqlx::{PgPool, Row};
